test-integration-pipeline/requirements.txt
pyspark==3.3.0
pandas==1.5.0
pyarrow==9.0.0
psycopg2-binary==2.9.5
flask==2.2.2
prometheus-client==0.15.0
pytest==7.2.0
requests==2.28.1
